Tournament On the Fly Changes
Completed the following:
Added asterisks to players in attendance (and full player list) currently in active matches
Added an undo button for a single match - TODO: Create a history to extend this this feature (ie add an "effect" to the match object), also fix this feature - it broke.
Added sorting to the Create match screen. TODO: Add a scroll bar for player lists.
Added double click features to active matches as a shortcut to record the match, and to the rankings list to the the selected player's current rating.
Adjusted some colors - further adjustment is needed.